Hitmonitor_2.1
完成击打检测滤波功能

+++ Software/TASK/controltask/controltask.h
@@ -6,8 +6,8 @@ extern u32 system_time;
 extern u16 weight[10];
 extern u8 pointer;
 extern u8 hitflag ;
-extern u8 hitcount;
-extern u8 ledoff_time;
+extern u16 hitcount;
+extern int ledoff_time;
 void control_task(void);
 void hitmonitor(void);
 #endif
